Move telnet/transport.py to telnet/telnet.py.
