+++ vumi/transports/telnet/telnet.py
