+++ vumi/transports/telnet/tests/test_telnet.py
